fixing warnings and refactoring
